포스팅과정 별점 float에서 double로 변경

+++ app/src/main/java/com/example/front_ui/DataModel/PostingInfo.java
@@ -20,12 +20,12 @@ public class PostingInfo implements Serializable {
     public String storeId;
     public String writerId;
     public float aver_star;
-    public List<Float> detail_aver_star; //{맛, 가성비, 서비스, 분위기}
+    public List<Double> detail_aver_star; //{맛, 가성비, 서비스, 분위기}
 
     public PostingInfo() {
     }
 
-    public PostingInfo(String storeName, String imagePathInStorage, int numLike, Object postingTime, String title, Map<String, Boolean> tag, String description, String writerName, String address, String storeId, String writerId, float aver_star, List<Float> detail_aver_star) {
+    public PostingInfo(String storeName, String imagePathInStorage, int numLike, Object postingTime, String title, Map<String, Boolean> tag, String description, String writerName, String address, String storeId, String writerId, float aver_star, List<Double> detail_aver_star) {
         this.storeName = storeName;
         this.imagePathInStorage = imagePathInStorage;
         this.numLike = numLike;
@@ -139,11 +139,11 @@ public class PostingInfo implements Serializable {
         this.aver_star = aver_star;
     }
 
-    public List<Float> getDetail_aver_star() {
+    public List<Double> getDetail_aver_star() {
         return detail_aver_star;
     }
 
-    public void setDetail_aver_star(List<Float> detail_aver_star) {
+    public void setDetail_aver_star(List<Double> detail_aver_star) {
         this.detail_aver_star = detail_aver_star;
     }
 }
