+++ app/src/main/java/com/example/front_ui/DataModel/StoreInfo.java
@@ -12,9 +12,9 @@ import java.util.List;
 public class StoreInfo{
     public String name;
     public String storeId;
-    public float aver_star;
+    public double aver_star;
     public String address;
-    public List<Float> detail_aver_star;
+    public List<Double> detail_aver_star;
     private ArrayList<PostingInfo> allItemsInSection;
     public GeoPoint geoPoint;
 
@@ -26,7 +26,7 @@ public class StoreInfo{
     public StoreInfo(String name,
                      float aver_star,
                      String address,
-                     List<Float> detail_aver_star,
+                     List<Double> detail_aver_star,
                      GeoPoint geoPoint
                      ) {
         this.name = name;
@@ -43,7 +43,7 @@ public class StoreInfo{
         return name;
     }
 
-    public float getAver_star() {
+    public double getAver_star() {
         return aver_star;
     }
 
@@ -51,7 +51,7 @@ public class StoreInfo{
         return address;
     }
 
-    public List<Float> getDetail_aver_star() {
+    public List<Double> getDetail_aver_star() {
         return detail_aver_star;
     }
